+++ gulpfile.js
@@ -1,20 +1,18 @@
-const { dest, series, src } = require("gulp");
-const babel = require("gulp-babel");
-const gulpClean = require("gulp-clean");
-const sourcemaps = require("gulp-sourcemaps");
-const uglify = require("gulp-uglify");
+const { dest, series, src } = require('gulp');
+const babel = require('gulp-babel');
+const gulpClean = require('gulp-clean');
+const sourcemaps = require('gulp-sourcemaps');
 
 function clean() {
-  return src("dist", { read: false, allowEmpty: true }).pipe(gulpClean());
+  return src('dist', { read: false, allowEmpty: true }).pipe(gulpClean());
 }
 
 function build() {
-  return src(["lib/**/*.js", "!**/test/", "!**/test/**"])
+  return src(['lib/**/*.js', '!**/test/', '!**/test/**'])
     .pipe(sourcemaps.init())
-    .pipe(babel({ presets: ["@babel/preset-env"] }))
-    .pipe(uglify())
-    .pipe(sourcemaps.write("."))
-    .pipe(dest("dist"));
+    .pipe(babel({ presets: ['@babel/preset-env', 'minify'] }))
+    .pipe(sourcemaps.write('.', { sourceRoot: '/lib' }))
+    .pipe(dest('dist'));
 }
 
 exports.clean = clean;