+++ ReleaseNotes.md
@@ -1,5 +1,10 @@
 ## RELEASE NOTES
 
+### Version 7.6.4
+
+- Add source mappings for minified dist code
+- Add babel-preset-minify, remove gulp-uglify
+
 ### Version 7.6.3
 
 - Fix for util.uid() to always return string value for compatibility with STRATO 5.2+